+++ core/class/Abeille.class.php
@@ -634,7 +634,7 @@ if (0) {
         }
 
         /* Check & update configuration DB if required. */
-        $dbVersion = config::byKey('DbVersion', 'Abeille', '');
+        $dbVersion = config::byKey('ab::dbVersion', 'Abeille', '');
         $dbVersionLast = 20220421;
         if (($dbVersion == '') || (intval($dbVersion) < $dbVersionLast)) {
             log::add('Abeille', 'debug', 'deamon_start_cleanup(): DB config v'.$dbVersion.' < v'.$dbVersionLast.' => Update required.');